+++ src/App.tsx
@@ -7,33 +7,25 @@ import {
 import "./App.css";
 import Login from "./components/Login";
 import Home from "./components/Home";
-import { useState } from "react";
+import { useSelector } from "react-redux";
+import { RootState } from "./store";
 
 function App() {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-  const handleLogin = () => {
-    setIsLoggedIn(true);
-  };
+  const isAuthenticated = useSelector(
+    (state: RootState) => state.auth.isAuthenticated,
+  );
 
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-100">
-      -
       <Router>
         <Routes>
           <Route
             path="/login"
-            element={
-              isLoggedIn ? (
-                <Navigate to="/home" />
-              ) : (
-                <Login onLogin={handleLogin} />
-              )
-            }
+            element={isAuthenticated ? <Navigate to="/home" /> : <Login />}
           />
           <Route
             path="/home"
-            element={isLoggedIn ? <Home /> : <Navigate to="/login" />}
+            element={isAuthenticated ? <Home /> : <Navigate to="/login" />}
           />
           <Route path="/" element={<Navigate to="/login" />} />
         </Routes>